% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Statial.R
\name{buildModelsByCellType}
\alias{buildModelsByCellType}
\title{Initiate Model Building By Cell Type}
\usage{
buildModelsByCellType(
  x,
  markers,
  type,
  covariates,
  method,
  isMixed,
  verbose,
  timeout
)
}
\arguments{
\item{type}{prefix to build models on}

\item{method}{}

\item{nCores}{}
}
\description{
Initiate Model Building By Cell Type.
}
\examples{
XYZ

}
