% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Statial.R
\name{calculateCellDistances}
\alias{calculateCellDistances}
\title{Calculates Pairwise Distance Between Cell Types Per Image}
\usage{
calculateCellDistances(singleCellData, nCores = 1, Rs = c(200))
}
\arguments{
\item{Rs}{}
}
\description{
Calculates distance from each cell to the nearest cell of every type per image
}
\examples{
XYZ

}
