% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Statial.R
\name{calculateInteractionModels}
\alias{calculateInteractionModels}
\title{Calculates Linear Models For Pairwise Interactions}
\usage{
calculateInteractionModels(
  singleCellData,
  markers,
  type = "dist200_",
  covariates = NULL,
  method = "lm",
  isMixed = FALSE,
  verbose = TRUE,
  timeout = 10,
  nCores = 1
)
}
\arguments{
\item{type}{prefix to build models on}

\item{method}{}

\item{nCores}{}
}
\description{
Calculates linear models For pairwise interactions with options of mixed models and robust regression.
}
\examples{
XYZ

}
