% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Statial.R
\name{calculateK}
\alias{calculateK}
\title{Calculates Imhomogenous K Function Between Cell Types}
\usage{
calculateK(singleCellData, nCores = 1, Rs = c(200))
}
\arguments{
\item{maxRS}{}
}
\description{
Calculates Imhomogenous K Function Between Cell Types by Image
}
\examples{
XYZ

}
