% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Statial.R
\name{calculateModelsAllInteractions}
\alias{calculateModelsAllInteractions}
\title{Calculates Linear Models For Pairwise Interactions for all Variations of the Type Argument}
\usage{
calculateModelsAllInteractions(
  singleCellData,
  markers,
  typeAll = c("dist"),
  covariates = NULL,
  method = "lm",
  isMixed = FALSE,
  verbose = FALSE,
  timeout = 10,
  nCores = 1
)
}
\arguments{
\item{typeAll}{prefix to build models on}

\item{method}{}

\item{nCores}{}
}
\description{
Calculates linear models For pairwise interactions for all variations of the type argument.
}
\examples{
XYZ

}
