% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Statial.R
\name{distanceCalculator}
\alias{distanceCalculator}
\title{Calculate Pairwise Distance Between Cell Types}
\usage{
distanceCalculator(singleCellData, maxRS = 200)
}
\arguments{
\item{maxRS}{}
}
\description{
Calculates distance from each cell to the nearest cell of every type
}
\examples{
XYZ

}
