% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exampleImage}
\alias{exampleImage}
\title{Example simulated image}
\format{
headData a SingleCellExperiment object
}
\usage{
exampleImage
}
\description{
This is a single image obtained from the simulateCompartment function using the following code:
}
\details{
set.seed(10)

#Simulating example image
simulation = simulateCompartment(includeTissue = FALSE)

#Selecting image where a significant conditional relationship exists
exampleImage = simulation$sig
}
\keyword{datasets}
