% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Statial.R
\name{fitInteractionModels}
\alias{fitInteractionModels}
\title{Formula Builder and Model Runner}
\usage{
fitInteractionModels(x, f, covariates, method, isMixed, timeout = timeout)
}
\arguments{
\item{method}{}

\item{timeout}{}

\item{type}{prefix to build models on}
}
\description{
Formula Builder and Model Runner
}
\examples{
XYZ

}
