% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Statial.R
\name{getDistances}
\alias{getDistances}
\title{Wrapper to calculate pairwise distance between cell types by image}
\usage{
getDistances(singleCellData, Rs = c(200), whichCellTypes = NULL, nCores = 1)
}
\arguments{
\item{singleCellData}{A dataframe with a cellType column as well as x and y spatial coordinates.
The dataframe must contain a imageID column and cellID (unique cell
identifier's) column as well}

\item{Rs}{Radius to calculate pairwise distances between cells (can be a numeric or
vector of radii)}

\item{whichCellTypes}{Character vector specifying what cell types to include in the calculation.
If the argument is non-null, then at least two cell types need to be
specified}

\item{nCores}{Number of cores for parallel processing}
}
\description{
Calculates the euclidean distance from each cell to the nearest cell of each
type
}
\examples{
library(dplyr)
data("headSCE")

singleCellDataDistances <- getDistances(headSCE,
  nCores = 1,
  Rs = c(200),
  whichCellTypes = c("MC2", "SC7")
)

}
