% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Statial.R
\name{getMarkerMeans}
\alias{getMarkerMeans}
\title{Extract the average expression for all markers for each cell type in each 
region defined by lisaClust}
\usage{
getMarkerMeans(
  data,
  survivalData,
  imageID = NULL,
  cellType = NULL,
  region = NULL,
  markers = NULL
)
}
\arguments{
\item{data}{A SingleCellExperiment object with intensities data in the assays slot and
regions information in colData generated by lisaClust.}

\item{survivalData}{A string vector specifying name of the column specifying information on 
patient survival in colData.}

\item{imageID}{A string vector of imageIDs to specify for which images the marker mean
needs to be calculated for. If NULL, all images will be used.}

\item{cellType}{A string vector of cell types to specify which cell types the marker mean
needs to be calculated for. If NULL, all cell types will be used.}

\item{region}{A string vector of regions provided by lisaClust so specify for which
regions the marker mean will be calculated. If NULL, all regions will be 
used.}

\item{markers}{A string vector of markers that proxy a cell's state. If NULL, all markers 
will be used.}
}
\description{
Takes a SingleCellObject and outputs a dataframe in a convenient format for 
cross validation
}
\examples{
library(dplyr)
data(kerenSCE)

regionSCE <- lisaClust::lisaClust(kerenSCE, k = 5)

lisaClustOutput2 <- getMarkerMeans(regionSCE,
                    survivalData = "Survival_days_capped")
}
