% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parentCombinations.R
\name{getParentPhylo}
\alias{getParentPhylo}
\title{Extract parent and all children from a Phylo object}
\usage{
getParentPhylo(phylo_tree)
}
\arguments{
\item{phlyo_tree}{a phylo object or a treekoR result.}
}
\value{
A named list of parents and their respective children.
}
\description{
This function takes in a `phylo` object or a `treekoR` result from the 
\code{\link[treekoR]{getClusterTree}} function, and converts its into a 
named list of each and children to input into 
\code{\link[Statial]{parentCombinations}}.
}
