% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Statial.R
\name{getStateChangesFast}
\alias{getStateChangesFast}
\title{Wrapper function to quickly build ordinary linear models measuring state
changes on a image basis}
\usage{
getStateChangesFast(
  singleCellData,
  markers,
  type = "dist",
  covariates = NULL,
  removeColsThresh = 0.1,
  cellTypesToModel = NULL,
  nCores = 1
)
}
\arguments{
\item{singleCellData}{A dataframe with a imageID, cellType, and marker intensity column along
with covariates (e.g. distance or abundance of the nearest cell type) to
model cell state changes}

\item{markers}{A string list of markers that proxy a cell's state}

\item{type}{A prefix that appears on the column names of all cell state modelling
covariates. The default value is "dist"}

\item{covariates}{A list of additional covariates to be included in the model being built}

\item{removeColsThresh}{A threshold in which a modelling covariate column will be excluded in a
image model if missingness exceeds a certain proportion}

\item{cellTypesToModel}{A string vector of the names of cell types to model cell state changes in.
The default argument is NULL which models are cell types}

\item{nCores}{Number of cores for parallel processing}
}
\description{
Builds linear models measuring marker based state changes in a cell type
based of the proximity or abundance of another cell type. The function only
provides the option to build OLS models
}
\examples{
data("headSCE")
intensitiesData <- data.frame(t(
  SummarizedExperiment::assay(headSCE, "intensities")
))
spatialData <- data.frame(colData(headSCE))
markersToUse <- colnames(intensitiesData)
singleCellData <- cbind(
  spatialData[rownames(intensitiesData), ], intensitiesData
)
singleCellData <- singleCellData \%>\%
  mutate_at(markersToUse, function(x) ifelse(is.na(x), 0, x)) \%>\%
  mutate_if(is.factor, as.character)

singleCellDataDistances <- getDistances(singleCellData,
  nCores = 1,
  Rs = c(200),
  whichCellTypes = c("MC2", "SC7")
)

imageModelsFast <- getStateChangesFast(
  singleCellData = singleCellDataDistances,
  markers = markersToUse,
  type = c("dist200"),
  nCores = 1
)
}
