% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{headSCE}
\alias{headSCE}
\title{Head and neck cutaneous squamous cell carcinoma IMC data}
\format{
headData a SingleCellExperiment object
}
\usage{
headSCE
}
\description{
This is a imaging mass cytometry dataset from Ferguson et al 2022. The object contains 29 images.
}
\keyword{datasets}
