% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Statial.R
\name{imageModelsCVFormat}
\alias{imageModelsCVFormat}
\alias{visualiseImageRelationship}
\title{Convert Image Model Output to Cross-Validation Format}
\usage{
imageModelsCVFormat(
  imageModels,
  values_from = "tValue",
  removeColsThresh = 0.2,
  missingReplacement = 0
)

visualiseImageRelationship(
  singleCellData,
  imageID,
  mainCellType,
  interactingCellType,
  depedentMarker,
  sizeVariable = NULL,
  shape = NULL,
  interactive = TRUE,
  plotModelFit = TRUE,
  method = "lm",
  modelType = "dist200_"
)
}
\arguments{
\item{interactingCellType}{}

\item{modelType}{}

\item{replacementValue}{}
}
\description{
Convert image model output to cross-validation format.

Visualise Cell-Cell Marker Relationships.
}
\examples{
XYZ

XYZ

}
