% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Konditional.R
\name{isKonditional}
\alias{isKonditional}
\title{Test whether an object is a konditionalResult}
\usage{
isKonditional(konditionalResult)
}
\arguments{
\item{konditionalResult}{a object to test}
}
\description{
Test whether an object is a konditionalResult
}
\examples{
data = data.frame()
if(!isKonditional(data)) print("Not a konditionalResult")

}
