% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Konditional.R
\name{isKontextual}
\alias{isKontextual}
\title{Test whether an object is a kontextualResult}
\usage{
isKontextual(kontextualResult)
}
\arguments{
\item{kontextualResult}{a object to test}
}
\description{
Test whether an object is a kontextualResult
}
\examples{
data <- data.frame()
if (!isKontextual(data)) print("Not a kontextualResult")

}
