% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kerenImage}
\alias{kerenImage}
\title{MIBI-TOF Breast cancer image}
\format{
kerenImage an image from a SingleCellExperiment object
}
\usage{
data(kerenImage)
}
\description{
This is a single MIBI-TOF image of breast cancer from patient 6 of the
Keren et al 2018 dataset.
}
\references{
Keren, L., Bosse, M., Marquez, D., Angoshtari, R., Jain,
S., Varma, S., Yang, S. R., Kurian, A., Van Valen, D., West, R., Bendall,
S. C., & Angelo, M. (2018). A Structured Tumor-Immune Microenvironment in
Triple Negative Breast Cancer Revealed by Multiplexed Ion Beam Imaging.
Cell, 174(6), 1373-1387.e1319. ([DOI](https://doi.org/10.1016/j.cell.2018.08.039))
}
\keyword{datasets}
