% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kerenKontextual}
\alias{kerenKontextual}
\title{Kontextual results from kerenSCE}
\format{
kerenKontextual a kontextual results object.
}
\usage{
data(kerenKontextual)
}
\description{
This is a kontextual results data.frame created using Kontextual on the kerenSCE dataset.
}
\keyword{datasets}
