% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Konditional.R
\name{Konditional}
\alias{Konditional}
\title{Identify changes in cell state of a cell type as it becomes closer to another, conditional on a third population.}
\usage{
Konditional(
  imageData,
  parentDf,
  r,
  from = NULL,
  to = NULL,
  parent = NULL,
  inhom = TRUE,
  edgeCorrect = FALSE,
  window = "convex",
  window.length = NA,
  weightQuantile = 0.8,
  includeZeroCells = TRUE,
  includeOriginal = TRUE,
  cores = 1
)
}
\arguments{
\item{imageData}{A SingleCellExperiment or SpatialExperiment or a list of single images.}

\item{r}{The radius or radii which pairwise cell relationships are evaluated at.}

\item{from}{= NULL,}

\item{to}{= NULL,}

\item{parent}{= NULL,}

\item{inhom}{A logical value indicating whether to perform an inhomogeneous L function.}

\item{edgeCorrect}{A logical value indicating whether to perform edge correction.}

\item{window}{= "convex",}

\item{window.length}{= NA,}

\item{weightQuantile}{= .80,}

\item{includeOriginal}{A logical value to return the original L function values along with the konditional values.}

\item{cores}{Number of cores for parallel processing.}
}
\value{
A Koditional result object
}
\description{
Identify changes in cell state of a cell type as it becomes closer to another, conditional on a third population.
}
\examples{

1+1

}
