% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Konditional.R
\name{KonditionalCore}
\alias{KonditionalCore}
\title{Core function used by Konditional}
\usage{
KonditionalCore(
  image,
  r,
  from,
  to,
  parent,
  inhom = TRUE,
  edge = FALSE,
  includeOriginal = TRUE,
  weightQuantile = 0.8,
  ...
)
}
\arguments{
\item{image}{A single image from a Single Cell Experiment object.}

\item{r}{The radius which pairwise cell relationships are evaluated at.}

\item{from}{The first cell type to be evaluated in the pairwise relationship.}

\item{to}{The second cell type to be evaluated in the pairwise relationship.}

\item{parent}{The parent population of the from cell type (must include from cell type).}

\item{inhom}{A logical value indicating whether to perform an inhomogeneous L function.}

\item{edge}{A logical value indicating whether to perform edge correction.}

\item{includeOriginal}{A logical value to return the original L function values along with the konditional values.}

\item{...}{Any arguments passed into \code{\link[Statial]{inhomLParent}}.}
}
\value{
A single row of a data frame containing the konditional and orignal L values.
}
\description{
Core function used by Konditional
}
\examples{
XYZ

}
