% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeWindow.R
\name{makeWindow}
\alias{makeWindow}
\title{Creates a window for PPP object}
\usage{
makeWindow(data, window = "square", window.length = NULL)
}
\arguments{
\item{window.length}{}
}
\description{
Creates a window for PPP object
}
\examples{
XYZ

}
