% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Statial.R
\name{modelsPerCell}
\alias{modelsPerCell}
\title{Formula Builder and Model Runner}
\usage{
modelsPerCell(x, markers, type, covariates, method, isMixed, verbose, timeout)
}
\arguments{
\item{type}{prefix to build models on}

\item{method}{}

\item{timeout}{}
}
\description{
Formula Builder and Model Runner
}
\examples{
XYZ

}
