% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Konditional.R
\name{prepKontextMat}
\alias{prepKontextMat}
\title{Convert Kontextual results to a matrix for classification}
\usage{
prepKontextMat(
  kontextualResult,
  type = "kontextual",
  replaceVal = 0,
  imageID = "imageID"
)
}
\arguments{
\item{kontextualResult}{a kontextual result data.frame}

\item{type}{return the either the `kontextual` or `original` L-function values}

\item{replaceVal}{value which NAs are replaced with}

\item{imageID}{The column which contains image identifiers.}
}
\description{
Convert Kontextual results to a matrix for classification
}
