% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Konditional.R
\name{prepMatrix}
\alias{prepMatrix}
\title{Convert Kontextual or state changes result to a matrix for classification}
\usage{
prepMatrix(result, replaceVal = 0, column = NULL, test = NULL)
}
\arguments{
\item{result}{a kontextual or state changes result data.frame.}

\item{replaceVal}{value which NAs are replaced with.}

\item{column}{The column which contains the scores that you want to select.}

\item{test}{A column containing which will be the column names of the expanded matrix.}
}
\description{
Convert Kontextual or state changes result to a matrix for classification
}
\examples{
data("kerenSCE")


CD4_Kontextual <- Kontextual(
  cells = kerenSCE,
  r = 50,
  from = "Macrophages",
  to = "Keratin_Tumour",
  parent = c("Macrophages", "CD4_Cell"),
  image = "6"
)


kontextMat <- prepMatrix(CD4_Kontextual)

}
