% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relabelKonditional.R
\name{relabelKonditional}
\alias{relabelKonditional}
\alias{relabel}
\title{Function which randomises specific cells in an image and cacluates the Konditional value.}
\usage{
relabelKonditional(
  image,
  nSim = 1,
  r,
  from,
  to,
  parent,
  inhom = TRUE,
  edge = FALSE,
  cores = 1,
  ...
)

relabel(image, labels = NULL)
}
\arguments{
\item{image}{A single image from a Single Cell Experiment object.}

\item{nSim}{Number of randomisations which will be calculated.}

\item{r}{Radius to evaluated pairwise relationships between from and to cells.}

\item{from}{The first cell type to be evaluated in the pairwise relationship.}

\item{to}{The second cell type to be evaluated in the pairwise relationship.}

\item{parent}{The parent population of the from cell type (must include from cell type).}

\item{inhom}{A logical value indicating whether to perform an inhomogeneous L function.}

\item{edge}{A logical value indicating whether to perform edge correction.}

\item{cores}{Number of cores for parallel processing.}

\item{...}{Any arguments passed into Konditional}

\item{labels}{A vector of CellTypes to be randomised. If NULL all cells labels will be radomised.}
}
\value{
A data frame containing all images

A data frame containing all pairwise cell relationships and their corresponding parent
}
\description{
Function which randomises specific cells in an image and cacluates the Konditional value.

Function to randomise all specified cells labels in an image
}
