% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsCurve.R
\name{rsCurve}
\alias{rsCurve}
\title{Evaluation of Konditional over a range of radii.}
\usage{
rsCurve(
  cells,
  from,
  to,
  parent,
  rs = seq(10, 100, 10),
  inhom = TRUE,
  edge = FALSE,
  se = FALSE,
  nSim = 20,
  cores = 1,
  ...
)
}
\arguments{
\item{cells}{A single image from a SingleCellExperiment object}

\item{from}{The first cell type to be evaluated in the pairwise relationship.}

\item{to}{The second cell type to be evaluated in the pairwise relationship.}

\item{parent}{The parent population of the from cell type (must include from cell type).}

\item{rs}{A vector of radii to evaluate konditional over.}

\item{inhom}{A logical value indicating whether to perform an inhomogeneous L function.}

\item{edge}{A logical value indicating whether to perform edge correction.}

\item{se}{A logical value to indicate if the standard deviation of
konditional should be calculated to construct error bars.}

\item{nSim}{Number of randomisations to perform using \code{\link[Statial]{relabelKonditional}},
which will be used to calculated the SE.}

\item{cores}{Number of cores for parallel processing.}

\item{...}{Any arguments passed into \code{\link[Statial]{Konditional}}.}
}
\value{
A data frame of original L values and Konditional values evaluated
over a range of radii.
}
\description{
This function obtains `Kondtional` values over a range of radii, standard 
deviations for each value can be obtained using permutation for significance 
testing. To obtain estimates for standard deviations specify `se = TRUE`.
}
\examples{

data("kerenImage")

rsDf <- rsCurve(
  cells = kerenImage,
  from = "p53",
  to = "Immune",
  parent = c("p53", "Keratin+Tumour"),
  rs = seq(10, 510, 100),
  cores = 2
)

}
