% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateCompartment.R
\name{simulateCompartment}
\alias{simulateCompartment}
\title{Creates simulated image of tumour immune compartmentalisation}
\usage{
simulateCompartment(
  removal = 0.25,
  r = 0.1,
  sigma = 0.05,
  k = 40,
  mu = 50,
  childSigma = 1,
  includeTissue = TRUE
)
}
\arguments{
\item{includeTissue}{}
}
\value{
A list containing two images, insig: where there is no conditional relationship, and sig: where there is a conditional relationship
}
\description{
Creates simulated image of tumour immune compartmentalisation
}
\examples{
images = simulateCompartment()
plot(images$sig)

}
