% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stateChangesContam}
\alias{stateChangesContam}
\title{Results from getStateChanges with the mixed effects model when using
contamination as a covariate for kerenSCE.}
\format{
stateChangesContam a dataframe
}
\usage{
stateChangesContam
}
\description{
Results from the call:
SCE <- getDistances(kerenSCE, nCores = 20)
SCE <- getAbundances(SCE, nCores = 20)
SCE <- calcContamination(SCE)
}
\details{
stateChangesContam <- getStateChanges(
  singleCellData = SCE,
  Rs = c(200),
  typeAll = c("dist200", "abundance200"),
  covariates = c("rfMainCellProb"),
  method = "lm",
  isMixed = TRUE,
  nCores = 40)
}
\keyword{datasets}
