% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stateChangesFiltered}
\alias{stateChangesFiltered}
\alias{stateChanges}
\title{Results from getStateChanges with the mixed effects model for kerenSCE}
\format{
stateChanges a dataframe
}
\usage{
stateChangesFiltered
}
\description{
Results from the call:
SCE <- getDistances(kerenSCE, nCores = 20)
SCE <- getAbundances(SCE, nCores = 20)
}
\details{
stateChangesMixed <- getStateChanges(
  singleCellData = SCE,
  Rs = c(200),
  typeAll = c("dist200", "abundance200"),
  method = "lm",
  isMixed = FALSE,
  nCores = 40)
}
\keyword{datasets}
