% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stateChangesMixed}
\alias{stateChangesMixed}
\title{Results from getStateChanges with the mixed effects model for kerenSCE}
\format{
stateChangesMixed a dataframe
}
\usage{
stateChangesMixed
}
\description{
Results from the call:
SCE <- getDistances(kerenSCE, nCores = 20)
SCE <- getAbundances(SCE, nCores = 20)
}
\details{
stateChangesMixed <- getStateChanges(
  singleCellData = SCE,
  Rs = c(200),
  typeAll = c("dist200", "abundance200"),
  method = "lm",
  isMixed = TRUE,
  nCores = 40)
}
\keyword{datasets}
