% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Statial-package.R
\docType{package}
\name{Statial-package}
\alias{Statial}
\alias{Statial-package}
\title{Statial: A package to identify changes in cell state relative to spatial associations}
\description{
Statial is a suite of functions for identifying changes in cell state. The functionality provided by Statial provides robust quantification of cell type localisation which are invariant to changes in tissue structure. In addition to this Statial uncovers changes in marker expression associated with varying levels of localisation. These features can be used to explore how the structure and function of different cell types may be altered by the agents they are surrounded with.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://sydneybiox.github.io/Statial} \url{https://github.com/SydneyBioX/Statial/issues}
  \item \url{https://sydneybiox.github.io/Statial/}
  \item Report bugs at \url{https://github.com/SydneyBioX/Statial/issues}
}

}
\author{
\strong{Maintainer}: Farhan Ameen \email{fame2827@uni.sydney.edu.au}

Authors:
\itemize{
  \item Farhan Ameen \email{fame2827@uni.sydney.edu.au}
  \item Sourish Iyengar \email{sourish.iyengar@sydney.edu.au}
  \item Alex Qin \email{alex.qin@sydney.edu.au}
  \item Shila Ghazanfar \email{shila.ghazanfar@sydney.edu.au}
  \item Ellis Patrick \email{ellis.patrick@sydney.edu.au}
}

}
\keyword{internal}
