% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Statial.R
\name{distanceCalculator}
\alias{distanceCalculator}
\title{Calculate pairwise distance between cell types}
\usage{
distanceCalculator(data, maxDist = 200, distFun = "min")
}
\arguments{
\item{data}{the single cell data of interest}

\item{maxDist}{Maximum distance between pairs of points to be counted as close
pairs.}

\item{distFun}{How to merge duplicate entries.}
}
\description{
Calculates the euclidean distance from each cell to the nearest cell of each
type for a single image
}
