% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Statial.R
\name{distanceCalculator}
\alias{distanceCalculator}
\title{Calculate pairwise distance between cell types}
\usage{
distanceCalculator(singleCellData, maxRS = 200)
}
\arguments{
\item{singleCellData}{the single cell data of interest}

\item{maxRS}{Maximum distance between pairs of points to be counted as close
pairs.}
}
\description{
Calculates the euclidean distance from each cell to the nearest cell of each
type for a single image
}
