% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsCurve.R
\name{ggplotRs}
\alias{ggplotRs}
\title{Plotting the original and konditional L values over a range of radii.}
\usage{
ggplotRs(rsDf)
}
\arguments{
\item{rsDf}{A data frame from \code{\link[Statial]{rsCurve}}.}
}
\value{
A ggplotly object showing the original and konditional L function
 values over a range of radii
}
\description{
Plotting the original and konditional L values over a range of radii.
}
\examples{
data("exampleImage")

rsDf = rsCurve(exampleImage,
from = "cd8_t_cells",
to = "tumour_cells",
parent = c("cd8_t_cells", "t_cells"),
rs = seq(0.01, 0.10, 0.01),
cores = 40)

ggplotRs(rsDf)

}
