% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsCurve.R
\name{ggplotRs}
\alias{ggplotRs}
\title{Plotting the original and konditional L values over a range of radii.}
\usage{
ggplotRs(rsDf)
}
\arguments{
\item{rsDf}{A data frame from \code{\link[Statial]{rsCurve}}.}
}
\value{
A ggplotly object showing the original and konditional L function
 values over a range of radii
}
\description{
Plotting the original and konditional L values over a range of radii.
}
\examples{
data("kerenImage")

rsDf <- rsCurve(
  cells = kerenImage,
  from = "p53",
  to = "Immune",
  parent = c("p53", "Keratin+Tumour"),
  rs = seq(10, 510, 100),
  cores = 40
)

ggplotRs(rsDf)

}
