% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsCurve.R
\name{ggplotRs}
\alias{ggplotRs}
\title{Plotting the original and konditional L values over a range of radii.}
\usage{
ggplotRs(rsDf, plot = TRUE)
}
\arguments{
\item{rsDf}{A data frame from \code{\link[Statial]{rsCurve}}.}

\item{plot}{A logical which will output a ggplotly object if `TRUE` or return a ggplot object if `FALSE`}
}
\value{
A ggplotly object showing the original and konditional L function values over a range of radii
}
\description{
Plotting the original and konditional L values over a range of radii.
}
\examples{
XYZ

}
