% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Statial.R
\name{listImageModelsCVFormat}
\alias{listImageModelsCVFormat}
\title{Convert Image Model Output to Cross-Validation Format}
\usage{
listImageModelsCVFormat(
  imageModels,
  classificationData,
  values_from = "tValue",
  removeColsThresh = 0.2,
  missingReplacement = 0
)
}
\arguments{
\item{imageModels}{A dataframe with the output from the function getStateChanges or
getStateChangesFast with the argument isMixed = FALSE}

\item{classificationData}{A level of factors with the classification conditions to be predicted 
and the imageID as names.}

\item{values_from}{Column to use from imageModels for cross validation. The default is
"tValue" but "beta" can be choosen as well}

\item{removeColsThresh}{Threshold of missingness in which a relationship will not be included as
column in the cross validation ready output}

\item{missingReplacement}{Numeric value to replace missing values}
}
\description{
Takes the output from the getStateChanges function for image based state
models and converts it in a convenient format for cross validation
}
\examples{
library(dplyr)
data("headSCE")

singleCellDataDistances <- getDistances(headSCE,
  nCores = 1,
  Rs = c(200),
  whichCellTypes = c("MC2", "SC7")
)

imageModels <- getStateChanges(
  singleCellData = singleCellDataDistances,
  Rs = c(200),
  type = c("dist200"),
  nCores = 1
)

classificationDataKeep <- factor(x = c("NP", "NP", "NP", "P", "P")) \%>\% 
purrr::set_names(c("1", "5", "6", "7", "8"))

crossValidationData <- listImageModelsCVFormat(imageModels,
  classificationData = classificationDataKeep,
  values_from = "tValue",
  removeColsThresh = 0.2,
  missingReplacement = 0
)
}
