% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Statial.R
\name{imageModelsCVFormat}
\alias{imageModelsCVFormat}
\title{Convert Image Model Output to Cross-Validation Format}
\usage{
imageModelsCVFormat(
  imageModels,
  values_from = "tValue",
  removeColsThresh = 0.2,
  missingReplacement = 0
)
}
\arguments{
\item{imageModels}{A dataframe with the output from the function getStateChanges or
getStateChangesFast with the argument isMixed = FALSE}

\item{values_from}{Column to use from imageModels for cross validation. The default is
"tValue" but "beta" can be choosen as well}

\item{removeColsThresh}{Threshold of missingness in which a relationship will not be included as
column in the cross validation ready output}

\item{replacementValue}{Numeric value to replace missing values}
}
\description{
Takes the output from the getStateChanges function for image based state
models and converts it in a convenient format for cross validation
}
\examples{
load("headSCE.rda")
intensitiesData <- data.frame(t(assay(headSCE, "intensities")))
spatialData <- data.frame(colData(headSCE))
markersToUse <- colnames(intensitiesData)
singleCellData <- cbind(
  spatialData[rownames(intensitiesData), ], intensitiesData
)
singleCellData <- singleCellData \%>\%
  mutate_at(markersToUse, function(x) ifelse(is.na(x), 0, x)) \%>\%
  mutate_if(is.factor, as.character)

singleCellDataDistances <- getDistances(singleCellData,
  nCores = 1,
  Rs = c(200),
  whichCellTypes = c("MC2", "SC7")
)

imageModelsFast <- getStateChangesFast(
  singleCellData = singleCellDataDistances,
  markers = markersToUse,
  type = c("dist200"),
  nCores = 1
)
crossValidationData <- imageModelsCVFormat(imageModelsFast,
  values_from = "tValue",
  removeColsThresh = 0.2,
  missingReplacement = 0
)
}
