% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inhomLParent.R
\name{inhomLParent}
\alias{inhomLParent}
\alias{inhomL}
\alias{PPPdf}
\alias{borderEdge}
\title{Calculates conditional L value}
\usage{
inhomLParent(
  data,
  Rs = 20,
  window = "convex",
  window.length = NULL,
  weightQuantile = 0.8,
  from = NULL,
  to = NULL,
  parent = NULL,
  inhom = TRUE,
  edgeCorrect = TRUE,
  includeZeroCells = TRUE,
  closePairs = NULL
)

inhomL(p, lam, X, Rs, num, Area)

PPPdf(ppp)

borderEdge(X, maxD)
}
\arguments{
\item{closePairs}{}

\item{Area}{}

\item{ppp}{a PPP object}

\item{maxD}{}
}
\description{
Calculates conditional L value

Calculates L value from weights and lambda values

Converts a PPP object to a data frame for inhomLParent function

Edge correction for L values
}
\examples{
XYZ

XYZ


XYZ


XYZ


}
