% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inhomLParent.R
\name{inhomLParent}
\alias{inhomLParent}
\title{Calculates conditional L value}
\usage{
inhomLParent(
  data,
  Rs = 20,
  window = "convex",
  window.length = NULL,
  weightQuantile = 0.8,
  from = NULL,
  to = NULL,
  parent = NULL,
  inhom = TRUE,
  edgeCorrect = TRUE,
  includeZeroCells = TRUE,
  closePairs = NULL
)
}
\arguments{
\item{data}{A single image from a Single Cell Experiment or point pattern object.}

\item{Rs}{The radius which pairwise cell relationships are evaluated at.}

\item{window}{Type of window for data, either `square`, `convex` or `concave`, passed into \code{\link[Statial]{makeWindow}}}

\item{window.length}{A tuning parameter for controlling the level of concavity when estimating concave windows.
Passed into \code{\link[Statial]{makeWindow}}}

\item{from}{The first cell type to be evaluated in the pairwise relationship.}

\item{to}{The second cell type to be evaluated in the pairwise relationship.}

\item{parent}{The parent population of the from cell type (must include from cell type).}

\item{inhom}{A logical value indicating whether to account for inhomogeneity.}

\item{edgeCorrect}{A logical value indicating whether to perform edge correction.}

\item{includeZeroCells}{A logical value indicating whether to include cells with}

\item{closePairs}{}
}
\description{
Calculates conditional L value
}
\examples{
XYZ

}
