% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Konditional.R
\name{Konditional}
\alias{Konditional}
\title{Identify changes in cell state of a cell type as it becomes closer to another, conditional on a third population.}
\usage{
Konditional(sce)
}
\arguments{
\item{sce}{A SingleCellExperiment}
}
\value{
A Statial result object
}
\description{
Identify changes in cell state of a cell type as it becomes closer to another, conditional on a third population.
}
\examples{

1+1

}
