% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Konditional.R
\name{Konditional}
\alias{Konditional}
\title{Evaluation of pairwise cell relationships, conditional on a 3rd population.}
\usage{
Konditional(
  cells,
  parentDf = NULL,
  r,
  from = NULL,
  to = NULL,
  parent = NULL,
  inhom = TRUE,
  edgeCorrect = FALSE,
  window = "convex",
  window.length = NA,
  weightQuantile = 0.8,
  includeZeroCells = TRUE,
  includeOriginal = TRUE,
  spatialCoords = c("x", "y"),
  cellType = "cellType",
  imageID = "imageID",
  cores = 1
)
}
\arguments{
\item{cells}{A SingleCellExperiment, SpatialExperiment or a list of
data.frames containing columns specifying the imageID, cellType, and x and y
spatial coordinates.}

\item{parentDf}{A data frame from \code{\link[Statial]{parentCombinations}}}

\item{r}{Radius to evaluated pairwise relationships between from and to cells.}

\item{from}{The first cell type to be evaluated in the pairwise relationship.}

\item{to}{The second cell type to be evaluated in the pairwise relationship.}

\item{parent}{The parent population of the from cell type (must include from cell type).}

\item{inhom}{A logical value indicating whether to account for inhomogeneity.}

\item{edgeCorrect}{A logical value indicating whether to perform edge correction.}

\item{window}{Type of window for data, either `square`, `convex` or `concave`,
passed into \code{\link[Statial]{makeWindow}}}

\item{window.length}{A tuning parameter for controlling the level of concavity when estimating concave windows.
Passed into \code{\link[Statial]{makeWindow}}}

\item{weightQuantile}{A decimal value indicating what quantile of parent
density used to weight the `from` cells.}

\item{includeZeroCells}{A logical value indicating whether to include cells
with zero counts in the pairwise association calculation.}

\item{includeOriginal}{A logical value to return the original L function
values along with the konditional values.}

\item{spatialCoords}{The columns which contain the x and y spatial coordinates.}

\item{cellType}{The column which contains the cell types.}

\item{imageID}{The column which contains image identifiers.}

\item{cores}{Number of cores for parallel processing.}
}
\value{
A konditionalResult object
}
\description{
Konditional identifies the relationship between two cell types which are 
conditional on the spatial behaviour of a 3rd cell population, for a 
particular radius (r).
}
\examples{
# Load data
data("headSCE")

CD4_Konditional <- Konditional(
  cells = headSCE,
  r = 50,
  from = "TC_CD4",
  to = "SC5",
  parent = c("TC_CD4", "TC_CD8"),
  cores = 40
)


head(CD4_Konditional)

}
