% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parentCombinations.R
\name{parentCombinations}
\alias{parentCombinations}
\title{Create all combinations of cell type relationships from a list of parents}
\usage{
parentCombinations(all, ...)
}
\arguments{
\item{all}{A list of all cell types}

\item{...}{Vectors of each parent population}
}
\value{
A data frame containing all pairwise cell relationships and their corresponding parent
}
\description{
Create all combinations of cell type relationships from a list of parents
}
\examples{
tcells = c("CD4", "CD8")
tissue = c("epithelial", "stromal")
allCells = c("tumour", tissue, tcells)

parentCombinations(all = allCells, tcells, tissue)


}
