% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Statial.R
\name{randomForestContaminationCalculator}
\alias{randomForestContaminationCalculator}
\title{Calculates Contamination Scores Based Using Random Forest}
\usage{
randomForestContaminationCalculator(
  singleCellData,
  markers,
  seed = 2022,
  num.trees = 100,
  verbose = FALSE,
  missingReplacement = 0
)
}
\arguments{
\item{missingReplacement}{}

\item{nCores}{}
}
\description{
Calculates contamination scores using random forest
}
\examples{
XYZ

}
