% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relabelKonditional.R
\name{relabelKonditional}
\alias{relabelKonditional}
\alias{relabel}
\title{Function which randomises specific cells in an image and cacluates the Konditional value.}
\usage{
relabelKonditional(
  image,
  nSim = 1,
  r,
  from,
  to,
  parent,
  returnImages = FALSE,
  inhom = TRUE,
  edge = FALSE,
  cores = 1,
  ...
)

relabel(image, labels = NULL)
}
\arguments{
\item{image}{A single image from a Single Cell Experiment object.}

\item{nSim}{Number of randomisations which will be calculated.}

\item{r}{Radius to evaluated pairwise relationships between from and to cells.}

\item{from}{The first cell type to be evaluated in the pairwise relationship.}

\item{to}{The second cell type to be evaluated in the pairwise relationship.}

\item{parent}{The parent population of the from cell type (must include from cell type).}

\item{returnImages}{A logical value to indicate whether the function should 
return the randomised images along with the Konditional values.}

\item{inhom}{A logical value indicating whether to account for inhomogeneity.}

\item{edge}{A logical value indicating whether to perform edge correction.}

\item{cores}{Number of cores for parallel processing.}

\item{...}{Any arguments passed into \code{\link[Statial]{Konditional}}}

\item{labels}{A vector of CellTypes labels to be permuted If NULL all cells 
labels will be radomised.}
}
\value{
A data frame containing Konditional value for each randomised image. 
If `returnImages = TRUE` function will return a list with Konditional values 
and the randomised images.

A data frame containing all pairwise cell relationships and their 
corresponding parent
}
\description{
Function which randomises specific cells in an image and cacluates the Konditional value.

Function to permute all specified cells labels in an image
}
\examples{
data("exampleImage")

relabelResult = relabelKonditional(
image = exampleImage,
nSim = 5,
r = 0.05,
from = "cd8_t_cells",
to = "tumour_cells",
parent = c("cd8_t_cells", "t_cells"),
cores = 40)

data("exampleImage")

#Permute CD8 T cells and T cell labels in the image
relabeledImage = relabel(exampleImage, labels = c("cd8_t_cells", "t_cells"))
plot(relabeledImage)

}
