% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relabelKonditional.R
\name{relabelKonditional}
\alias{relabelKonditional}
\alias{relabel}
\title{Cell permutation for Konditional}
\usage{
relabelKonditional(
  cells,
  nSim = 1,
  r,
  from,
  to,
  parent,
  returnImages = FALSE,
  inhom = TRUE,
  edge = FALSE,
  cores = 1,
  spatialCoords = c("x", "y"),
  cellType = "cellType",
  imageID = "imageID",
  ...
)

relabel(image, labels = NULL)
}
\arguments{
\item{cells}{A single image data frame from a SingleCellExperiment object}

\item{nSim}{Number of randomisations which will be calculated.}

\item{r}{Radius to evaluated pairwise relationships between from and to cells.}

\item{from}{The first cell type to be evaluated in the pairwise relationship.}

\item{to}{The second cell type to be evaluated in the pairwise relationship.}

\item{parent}{The parent population of the from cell type (must include from cell type).}

\item{returnImages}{A logical value to indicate whether the function should
return the randomised images along with the Konditional values.}

\item{inhom}{A logical value indicating whether to account for inhomogeneity.}

\item{edge}{A logical value indicating whether to perform edge correction.}

\item{cores}{Number of cores for parallel processing.}

\item{...}{Any arguments passed into \code{\link[Statial]{Konditional}}}

\item{image}{A single image from a Single Cell Experiment object.}

\item{labels}{A vector of CellTypes labels to be permuted If NULL all cells
labels will be radomised.}
}
\value{
A data frame containing Konditional value for each randomised image.
If `returnImages = TRUE` function will return a list with Konditional values
and the randomised images.

A data frame containing all pairwise cell relationships and their
corresponding parent
}
\description{
Function which randomises specified cells in an image and calculates
the `Konditional` value. This can be used to estimate the null distribution, 
of the parent cell population for significance testing.

This function relabels all specified cells within a single image, to 
estimate the null distribution of cell population specified.
}
\examples{
data("kerenImage")

relabelResult <- relabelKonditional(
  cells = kerenImage,
  nSim = 5,
  r = 250,
  from = "p53",
  to = "Immune",
  parent = c("p53", "Keratin+Tumour"),
  cores = 2
)

data("kerenImage")

# Permute CD8 T cells and T cell labels in the image
relabeledImage <- relabel(kerenImage, labels = c("p53", "Keratin+Tumour"))
plot(relabeledImage)

}
