% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Statial.R
\name{visualiseImageRelationship}
\alias{visualiseImageRelationship}
\title{Visualise Cell-Cell Marker Relationships}
\usage{
visualiseImageRelationship(
  data,
  imageID,
  mainCellType,
  interactingCellType,
  depedentMarker,
  sizeVariable = NULL,
  shape = NULL,
  interactive = FALSE,
  plotModelFit = FALSE,
  method = "lm",
  modelType = "dist200_"
)
}
\arguments{
\item{modelType}{}
}
\description{
Helper functions to visualise OLS model fits for image based state models
}
\examples{
data("headSCE")
intensitiesData <- data.frame(t(
  SummarizedExperiment::assay(headSCE, "intensities")
))
spatialData <- data.frame(colData(headSCE))
markersToUse <- colnames(intensitiesData)
singleCellData <- cbind(
  spatialData[rownames(intensitiesData), ], intensitiesData
)
singleCellData <- singleCellData \%>\%
  mutate_at(markersToUse, function(x) ifelse(is.na(x), 0, x)) \%>\%
  mutate_if(is.factor, as.character)

singleCellDataDistances <- getDistances(singleCellData,
  nCores = 1,
  Rs = c(200),
  whichCellTypes = c("MC2", "SC7")
)
visualiseImageRelationship(
  singleCellData = singleCellDataDistances,
  imageID = "36",
  mainCellType = "MC2",
  interactingCellType = "SC7",
  depedentMarker = "Podoplanin",
  interactive = FALSE,
  plotModelFit = FALSE,
  method = "lm",
  modelType = "dist200_"
)

}
